\name{coordinates}
\alias{coordinates}
\docType{data}
\title{
  Coordinates of the dataset "hirlam"
}
\description{

The \code{coordinates} data is a data set that contains the spatial coordinates of the space-time data \code{hirlam}.

}


\format{
  A data frame with 70 observations on the following 2 variables.
  \describe{
    \item{\code{Longitude}}{ the longitude coordinates. }
    \item{\code{Latitude}}{ the latitude coordinates.}
  }
}


\keyword{datasets}